[refactor] Project,Period Selector 공용 컴포넌트 이동
- 페이지에 공용으로 사용될 셀럭터 컴포넌트 위치 이동
- period selector에 전체 기간 추가
